requirements.txt
Flask
mysql-connector-python
razorpay